+++ src/v1/invites/mod.rs
@@ -3,21 +3,22 @@
 // Copyright: 2023–2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-pub mod models;
 pub mod routes;
 
-pub use models::*;
 pub use routes::*;
 
 use rocket::Route;
 
 pub(super) fn routes() -> Vec<Route> {
     routes![
-        get_members,
-        search_members,
-        get_member,
-        set_member_role,
-        set_member_mfa,
-        logout_member,
+        invite_member,
+        get_invites,
+        get_invite,
+        get_invite_by_token,
+        invite_accept,
+        invite_reject,
+        invite_resend,
+        invite_cancel,
+        cancel_invite,
     ]
 }
